More general restructuring and some new files added to Make commands.
